+++ package.json
@@ -13,6 +13,9 @@
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.24.0",
+    "@react-native-firebase/app": "^20.3.0",
+    "@react-native-firebase/auth": "^20.3.0",
+    "@react-native-firebase/crashlytics": "^20.3.0",
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/stack": "^6.4.1",
@@ -21,7 +24,6 @@
     "expo-build-properties": "^0.12.5",
     "expo-dev-client": "~4.0.21",
     "expo-status-bar": "~1.12.1",
-    "firebase": "^10.12.5",
     "react": "18.2.0",
     "react-hook-form": "^7.52.2",
     "react-native": "0.74.3",
