+++ src/screens/home/Home.tsx
@@ -5,24 +5,9 @@ import { BottomTabBarButtonProps } from '@react-navigation/bottom-tabs'
 import { RootNavigationParam } from '../../app/navigation/types'
 
 export const Home = () => {
-  // const navigation = useNavigation<BottomTabBarButtonProps<RootNavigationParam, 'Profile'>>()
-
-  const goToProfile = () => {
-    //   navigation.navigate('Profile')
-  }
-
   return (
-    <View
-      style={{
-        display: 'flex',
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center',
-      }}
-    >
-      <TouchableOpacity onPress={goToProfile}>
-        <Text>go to profile</Text>
-      </TouchableOpacity>
+    <View>
+      <Text>Home</Text>
     </View>
   )
 }
